Security with JWT security Configurations

+++ src/main/java/firsttask/taskmanager/Controller/TaskController.java
@@ -1,6 +1,6 @@
 package firsttask.taskmanager.Controller;
 
-import firsttask.taskmanager.Exceptions.GeneralException;
+
 import firsttask.taskmanager.Exceptions.UserNotFoundException;
 import firsttask.taskmanager.Repositories.TaskRepository;
 import firsttask.taskmanager.Repositories.UserRepository;
@@ -8,11 +8,14 @@ import firsttask.taskmanager.domain.Task;
 import firsttask.taskmanager.domain.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.nio.file.AccessDeniedException;
 import java.util.List;
+import java.util.Optional;
 
 
 @RestController// don't ever forget this annotation
@@ -31,7 +34,7 @@ public class TaskController {
 
     //return all tasks
     @GetMapping("/tasks")
-    public List<Task> returnAllTasks() throws GeneralException{
+    public List<Task> returnAllTasks() throws Exception{
         logger.info("A get all tasks  request initialized ");
         List<Task> tasks = taskRepository.findAll();
         logger.trace("retrieve all tasks ");
@@ -42,27 +45,35 @@ public class TaskController {
 
     //return a task by its id
     @GetMapping("/tasks/{id}")
-    public Task returnTask(@PathVariable Long id) throws GeneralException{
+    public Task returnTask(@PathVariable Long id) throws Exception, AccessDeniedException {
         logger.info("A get task request initialized ");
+
         Task task = taskRepository.findById(id) //
                 .orElseThrow(() -> new UserNotFoundException(id));
-        logger.trace("retrieve task with id "+ id );
-        return  task;
+        User requestedUser=task.getUser();
+        User requestingUser= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (requestedUser.getId()==requestingUser.getId() && requestedUser.getPassword().equals(requestingUser.getPassword())) {
+
+            logger.trace("retrieve task with id "+ id );
+        return  task;}
+        else {
+            throw new AccessDeniedException("You are not allowed to access this page!");
+        }
     }
 
 
 
     @PostMapping("/tasks")
-    Task CreateTask(@RequestBody Task task, @RequestParam Long id ) throws GeneralException{
+    Task createTask(@RequestBody Task task) throws Exception{
 
         logger.info("A create task request initialized ");
-        User user = userRepository.findById(id).orElseThrow(()->new UserNotFoundException(id));
 
+        User requestingUser= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
             Task newTask = taskRepository.save(task);
           //  task.setUser(user);
             //notice that we have to add the task to the user object and add the user object to the task so that jpa can create load the table correctly for us
-            user.addTask(task);
-            userRepository.save(user);
+            requestingUser.addTask(task);
+            userRepository.save(requestingUser);
             logger.trace("Creating new  task"+newTask);
             return newTask;
 
@@ -72,15 +83,22 @@ public class TaskController {
 
     // edit the task from the owner user only so we make authentication first
     @PutMapping("/tasks/{id}")
-    Task edTask(@RequestBody Task editTask, @PathVariable Long id) throws GeneralException{
+    Task edTask(@RequestBody Task editTask, @PathVariable Long id) throws Exception, AccessDeniedException {
         logger.info("A Update task request initialized ");
-        Task updatedTask= taskRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
+        Task task = taskRepository.findById(id) //
+                .orElseThrow(() -> new UserNotFoundException(id));
+        User requestedUser=task.getUser();
+        User requestingUser= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (requestedUser.getId()==requestingUser.getId() && requestedUser.getPassword().equals(requestingUser.getPassword())) {
+
+            Task updatedTask = taskRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
             updatedTask.setDescription(editTask.getDescription());
             updatedTask.setCompleted(editTask.isCompleted());
-            // the correct way is that we must give a request param for each of the attributes and change what required
-            logger.trace("Updating a task to a user with id : "+ id +" The task : "+updatedTask);
+            logger.trace("Updating a task to a user with id : " + id + " The task : " + updatedTask);
             return updatedTask;
-
+        } else {
+            throw new AccessDeniedException("You are not allowed to access this page!");
+        }
 
     }
 
@@ -88,18 +106,25 @@ public class TaskController {
 
     // delete a task from the owner only
     @DeleteMapping("/tasks/{id}")
-    void  deleteTask(@PathVariable Long id,  HttpServletResponse response) throws IOException,GeneralException {
+    void  deleteTask(@PathVariable Long id,  HttpServletResponse response) throws IOException,Exception {
         logger.info("A delete task  request initialized ");
+        Task task = taskRepository.findById(id) //
+                .orElseThrow(() -> new UserNotFoundException(id));
+        User requestedUser=task.getUser();
+        User requestingUser= (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (requestedUser.getId()==requestingUser.getId() && requestedUser.getPassword().equals(requestingUser.getPassword())) {
 
-        if (taskRepository.existsById(id)) {
-            taskRepository.deleteById(id);
-        }
+            if (taskRepository.existsById(id)) {
+                taskRepository.deleteById(id);
+            }
 
 
-            logger.trace("Redirecting to the Tasks page after deleting task with id : " +id);
+            logger.trace("Redirecting to the Tasks page after deleting task with id : " + id);
             response.sendRedirect("");
 
-
+        }else {
+            throw new AccessDeniedException("You are not allowed to access this page!");
+        }
 
     }
